refactoring arena module
